fix URL parsing in serveSinglePageApp
this fixes an issue in serveSinglePageApp where the raw `request.url` string is being used as part of a conditional to retrieve static content - if a query parameter is set, it breaks the URL lookup. credit to @sgiacosa who wrote this code in #72 - this PR is a _very slight_ adaptation of his code sample!

closes #72