Support ES Modules.
This PR provides a possible solution for getting Workers Sites working with ES Module workers. This approach is not as invasive as other approaches.

Usage:

```
import manifestJSON from '__STATIC_CONTENT_MANIFEST'
const manifest = JSON.parse(manifestJSON)

export default {
  fetch(request, env, ctx){
    return await getAssetFromKV({request, waitUntil: ctx.waitUntil}, {
      ASSET_NAMESPACE: env.ASSET_NAMESPACE,
      ASSET_MANIFEST: manifest
    })
    // ...
  }
}
```

I'll also send a PR to wrangler that adds the manifest as a text module to the worker.