Hiding bash/zsh completion option.
It 'works', but has to be polished to be made really good. Of note, I can't figure out how to let it include file system options. Further, it only really works when wrangler's installed as a global, which we're moving away from.

This commit leaves it as working, but hides it from the menu.