Pick correct config object when passing env argument (#61)
* Pick correct config object when passing env argument

When you start your project with `npx wrangler@beta dev src/index.ts --env staging` it would fail with this error:

```
/Users/luzanne/Projects/my-worker/node_modules/wrangler/wrangler-dist/cli.js:27069
            throw ex;
            ^

TypeError: Cannot read properties of undefined (reading 'jsxFactory')
```

After debugging some stuff I found out that the `envRootObj` was not set correctly.`config['env.staging']` does not work, accessing properties goes only 1 deep.

This PR will fix that issue and use the correct config provided by the env variable.

Might be better to also add an extra check here. When an env variable is provided in the arguments, it must be available in the config as well? Thats up to you :)

Config object:
```
{
  _: [ 'dev' ],
  env: 'staging',
  local: false,
  ip: '127.0.0.1',
  port: 8787,
  'local-protocol': 'http',
  localProtocol: 'http',
  'upstream-protocol': 'https',
  upstreamProtocol: 'https',
  '$0': 'wrangler',
  filename: 'src/index.ts',
  config: {
    name: 'my-worker',
    type: 'javascript',
    zone_id: 'xxx',
    account_id: 'xxx',
    workers_dev: true,
    env: { staging: [Object], production: [Object] },
    build: { command: 'npm i && npm run build', upload: [Object] },
    __path__: '/Users/luzanne/Projects/my-worker/wrangler.toml'
  }
}
```

* Update index.tsx

Default an env object to {} if it isn't defined in the config.

* actually read envs correctly

* stray console.log

Co-authored-by: Sunil Pai <threepointone@gmail.com>
Co-authored-by: Sunil Pai <spai@cloudflare.com>