Implement `--polyfill-node`/ `config.polyfill_node`
This commit adds `--polyfill-node` as a cli argument and `polyfill_node` as a config field. It uses https://github.com/remorses/esbuild-plugins / https://github.com/ionic-team/rollup-plugin-node-polyfills/ for the polyfills. I also added a warning for the tradeoffs whenever this config is used. We may want better polyfills in the future, and maybe even better messaging if we can recognise popular libraries and suggest replacements.