don't log file contents when writing via `kv:key put <key> --path <path>` (#111)
* don't log file contents when writing via `kv:key put <key> --path <path>`

There are a couple of reasons we don't want to log the contents; it could be secret, or it could just be very long. This commit changes the messaging to instead just say which file is being uploaded. I also took this opportunity to add a log when we're deleting a key with `kv:key delete <key>`.

* Create clever-toys-retire.md