enable @typescript-eslint/no-floating-promises, pass lint+type check
Enabling that lint rule exposed a bunch of places where we weren't logging errors, and also caught at least 3 bugs. Glad I did this.

I also disabled prettier as an eslint error,  and added it to the `check` script instead.

I added `ioredis` as a root dependency. We don't use it, but we can't pass typecheck without it installed (because miniflare uses it, even tho it's an optional dep, and there's no good way to disable a third party's type errors in this fashion)

I also added `react-error-boundary` so we don't blow up completely, and have an option to handle errors from `dev` gracefully.