test: use a temporary directory for test files
This allows us to have more control over the directory structure under test.

For example, currently there is a package.json and a tsconfig.json above the test directory.
We use `findUp()` to check for these, so it is not possible (without mocking findUp) to test cases where these do not exist.