test: sort out the error logging
The way that yargs handles errors is problematic.
It was set up to catch errors in handlers; print the help info, along with the error message; and then re-throw the error.

But it was difficult to distinguish between errors that needed the help info, and those that didn't.
As a result, there were a number of places where we just printed to console.error and then returned. In these cases we could not exit the process with a non-zero error code.

Now, we have taken back control (!) of the error handling.
Further a number of Error types have been provided to self-document what is going on:

- Error: standard errors are just problems with something in the processing of the command; perhaps a network failure etc.
- CommandLineArgsError - used when there is a problem with the command line arguments, resulting in the help info being printed.
- DeprecationError - used just as a flag that this error is due to something being deprecated.
- NotImplementedError - used just as a flag that this error is due to a feature that has not yet been implemented.

Any error that gets to the top level will result in the process exiting with a code of 1.