pass bindings correctly to miniflare in `dev`
When calling miniflare in `dev` the bindings are passed into the process that's spawned incorrectly; each arg needs to be a separate element in the args array to the spawn call. This PR fixes how they're added to the args array. It also enables logs on the miniflare process. (they're a bit noisy, but we can make that better separately.)

I'm personally annoyed because I can swear this was working previously, but without tests that's just my word against... my own word, lol.

fixes #172