fix: prevent `dev` from infinitely restarting
In https://github.com/cloudflare/wrangler2/pull/164/, we added `token` to `useWorker`'s `useEffect` array that detects when to start the server process. However, the effect itself sets `token`, so the build would inifinitely restart on every run.

We could potentially fix this by using `useMemo` or something, but that's overkill for the reason we use `token` - to detect whether it's the first or subsequent time we've started the server, and what message to log to the terminal. So we introduce a ref `startedRef` to play that role, and mark it as true when we log the message the first time. This fixes the infinite loop.

This bug shows how urgent it is to write tests for everything else. We'll get there.