update-package-lock.json
Whenever we do a release, we have a problem where package-lock.json isn't in sync anymore. Specifically, because workspace definitions are also stored in this file, and we don't update the version numbers, there's a mismatch.

I would've just committed this directly to `main`, but I'm opening the PR to make this problem visible, and see if anyone has any suggestions for hot to fix this. Could the "Version Packages" PR generation process also run `npm install` at the root so package-lock.json also gets updated? Is this something the `changesets` team needs to fix?