Refactor `index.tsx` (and `config.ts`) to pass strict-null checks
This actually slightly changes the processing of the environment configuration.
Previously there were corner cases where the configuration might just crash.
These are now handled more cleanly with more appropriate warnings.

The yargs commands have been refactored slightly to be more consistent.
Now each command that can be either in local or remote mode is structured the same way.
The `config` is extracted from the args, then there is an `if-then-else` statement that handles the local and remote cases.
This ensures that the `account_id` has the correct typings in remote mode as needed.