chore: add workers-chat-demo (#475)
This adds the `workers-chat-demo` from https://github.com/cloudflare/workers-chat-demo as an example app into our monorepo. It has some modifications, documted at the top of its `README.md`.

Of note, you'll note that this _doesn't_ work with wrangler. Landing this so we can debug and have a shared reference that we need to get to work.