ci: enable artifact comments in PR from forks
We store the wrangler build artifacts and then comment on the PR to tell users how to access them.
But the way that GH actions work, it is not possible to write a comment if the PR is from another fork.

This change uses a `workflow_run` GH action, which does have permission, to do the commenting.
We must pass the PR number to it via another artifact.