chore: better RequestEvent type for tail
This adds a slightly better type for the `RequestEvent` used in `wrangler tail`. It's a "standard" type that `undici` uses, that's closer to the standard type we have in the Workers runtime.

I also deleted an unused function, I assume we'll bring that back when we write a test for tailing scheduled events.

Skipping a changeset here since there's no functional change.