feat: add support for `--local-protocol=https` to `wrangler dev`
This change adds full support for the setting the protocol that the localhost proxy server listens to.
Previously, it was only possible to use `HTTP`. But now you can set it to `HTTPS` as well.

To support `HTTPS`, Wrangler needs an SSL certificate.
Wrangler now generates a self-signed certificate, as needed, and caches it in the `~/.wrangler/local-cert` directory.
These certificates expire after 30 days and are regenerated by Wrangler as needed.

Note that if you use HTTPS then your browser will complain about the self-signed and you must tell it to accept the certificate before it will let you access the page.