Go back to using `getBindingNames`
So...a bunch of tests started failing after I rebased. Turns out,
we maybe can't trust the inputs to `validateBindingsHaveUniqueNames`
to be valid since we don't actually _throw_ until we've performed
every config validation. So that means by the time we get to
`validateBindingsHaveUniqueNames`, we might have some invalid stuff
in the `Config` and we can't really trust the types. :(

So I've just gone back to using `getBindingNames` to get the binding
names, since it does all the null & type checking necessary to get
the binding names out from a binding.