Get out in front of trademark issues with `wrangler jeans` and `wrangler jeep`
It's no secret that the wrangler 2 release is going to be big. Bigger than big. In fact, I don't think it's too much of a stretch to say it will change the way that software is developed and deployed on a huge scale. Like, on the same level as [git](https://blog.codacy.com/the-impact-of-git-on-software-development), or [CI/CD](https://jhall.io/archive/2021/09/26/a-brief-history-of-ci/cd/), or [debugging](https://revdebug.com/blog/debugging-history-origins/).

As a result, there are bound to be lawsuits incoming. Lawyers clamoring at Cloudflare's door with spurious claims like "You can't monopolize the entire serverless market with your ridiculously sleek, top-of-the-line, dev-friendly CLI!" and "We're suing for damages because our team was unable to work for a month straight after your project became so popular that it single-handedly took down GitHub."

These should be easily dealt with by Cloudflare's legal team. However, there is a potential issue regarding the naming of our CLI tool. Once wrangler becomes popular enough (which it will), and our SEO gets juiced enough (which it will), searches for "wrangler" will lead to our CLI first, and [jeans](https://www.wrangler.com/) and [Jeeps](https://www.jeep.com/wrangler.html) second.

This PR aims to address these complaints before they happen, by incorporating both of those products into our own. Embrace, Extend, Extinguish. Eventually, of course, it will be buying "Cloudflare Jeans" and "Jeep Cloudflares", but until the inevitable merger between the three companies it's probably best to throw a bone their way.