it runs now...but output isn't captured in tests...
sounds like a problem for the cass of tomorrow.

i also had to make some changes to the mocks for upload requests
because the filename isn't always `index.js`, i dunno if that's
a big deal or not.

Also there's probably a better way to check if they're doing sourcemaps,
because it looks like webpack generates a default thing.