Add support for publishing to Custom domains (#982)
* Support "custom_domain" boolean in toml

Needed to re-write a "renderer" of the route to be logged out to the
user, but it's now supported in the type system

* Publish Custom Domains via /domains api

If a toml provides routes with the "custom_domain" flag, publish those
separately from normal routes, using the /domains api. As these are more
complex than simple routes, we ask the api to error eagerly. If it
errors on a conflict for existing custom domains or managed DNS records,
we prompt for confirmation to override the conflicts, and then retry
with updated parameters

* Add tests for publishing to Custom Domains

* Add changeset